Add "holistic heuristic" algorithm for releases
This attempts to take into account EVRs in other Fedora versions, to
ensure that the new calculated release exceeds all previous builds in
the current or older Fedora versions, but doesn't do so with latest
builds in newer Fedora versions, if possible. It gives a little
wiggle-room to accommodate parallel builds of the (almost) same EVR
between different Fedora versions.

Signed-off-by: Nils Philippsen <nils@redhat.com>