Work on the logic to generate the changelog
In this commit, we are dropping the dependency on pygit2.
Fixes https://pagure.io/Fedora-Infra/rpmautospec/issue/10

We also take into account the presence and content of the
`changelog` file in the git repo as the idea is only to
automatically generate the changelog up until the latest commit
updating this `changelog` file.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>