Move all the BuildRequires higher in the spec and add the missing one on git
BuildRequires are for the entire spec file they do not depend on sub-packages
so it does not make sense to tie them with one.

We also need to have ``git`` installed in the buildroot for the test to run
successully, so we're adding it to the list of BuildRequires while at it.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>