Make tagging in Pagure generally available
Move the code from the hub plugin callback into `py2compat.tagging`,
into `PagureTaggingProxy`. Put it into a class to avoid having to supply
common parameters on every tagging operation.

In the course, fix the names of the tags being created and adapt/improve
how the hub plugin is tested.

Signed-off-by: Nils Philippsen <nils@redhat.com>