Run process-distgit in the build root if necessary
Steals the `broot` object from the BuildSRPMFromSCMTask.handler() caller
if it is not supplied to the callback as the parameter `buildroot`.

Adapt tests accordingly, mainly: because the bulk of the work is done
using the CLI in the build root, move testing it into the tests for
`rpmautospec.process_distgit`.

Signed-off-by: Nils Philippsen <nils@redhat.com>