Use logging instead of print()
This allows setting the log level globally, relieving us from having to
take care of it in the code.

Additionally, use %-style format strings for log messages to enable lazy
interpolation of format strings (and skip it altogether if the level
isn't handled).

Bump some log messages from INFO to DEBUG where this is better.

Signed-off-by: Nils Philippsen <nils@redhat.com>