Fix and improve release bumping algorithm
Previously, `evr` was overloaded, meaning both "latest existing version"
(mainly for testing) and "epoch-version for which we want a release". If
a new version was specified, this caused the latest existing release to
be assumed as "1", giving "2" as a result. This change separates these
two concerns.

Additionally, if nothing is specified the code looks into the spec file
for the [epoch:]version we need to work with.

Also, name some things more uniformly, e.g. use "next" consistently
instead of "new".

Signed-off-by: Nils Philippsen <nils@redhat.com>