Handle the situation where a commit has multiple git tags
To support this case git_get_tags must return a dict that has the hash
as key and a list of the tags associated with this commit as values.

Then the release process must handle the builds one by one, while the
changelog process will only take the most recent build into account.

In addition this commit also renames the variable `nevr` to `evr` to
reflect what it actually is.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>