Give concrete advice when adding changelog file
Multiple early adopters have encountered a problem
where an automatic changelog entry is not generated
for the commit that introduces the changelog file.
This happens because users attempt to introduce rpmautospec
together with other changes like a version update
and do not realize that the automatic changelog generation
happens only after the last commit touching the changelog file,
not starting from that commit.

Signed-off-by: Otto Urpelainen <oturpe@iki.fi>