doc: update docs to recommend "modern versioning"
The Packaging Guidelines have been adjusted a few months ago [1] to make use of
tilde and caret support in rpm, and in effect significantly simplified. The
version information about a package (including pre- and post-releases,
snapshots, and non-sortable upstream versions) is included in the Version field,
and the Release field is allowed to remain a simple counter (except in a few
special cases). This actually works great with %autorelease, since there's no
interference between the parts of Version that were embedded in the Release
field. So it's nice to be able to describe the simplified form of versioning
here too, since it makes %autorelease use significantly simpler.

This patch mirrors the change done in the Packaging Guidelines: a description of
"traditional versioning" is still present, but it's deemphasized and moved to a
separate chapter. It is still allowed and useful for RHEL7- systems, but the
simpler form is recommended.

Anchors were changed in the Packaging Guidelines rewrite, so they needed to be
adjusted.

v2:
- '-p' is also relegated to the section about "Traditional Versioning".

  This is correct thing for almost all packages, they shouldn't use '-p'.
  I'm aware of one exception: fedora-release uses "-p" before the final
  release. This is because fedora-release wants a version-release to be
  N-0.x and then N-y, e.g. 35-0.1, 35-0.2, 35-0.3, 35-1, 35-2, 35-3.
  But that is a very special case, and it actually does not work well. We
  currently get 35-0.1, 35-0.2, 35-0.3, 35-4, 35-5, 35-6, which is functional,
  but not pretty. In the end, it seems appropriate to "deprecate" "-p" here.
  The mechanism in fedora-release might need to be adjusted anyway.

[1] https://pagure.io/packaging-committee/pull-request/1073

Signed-off-by: Zbigniew Jędrzejewski-Szmek <zbyszek@in.waw.pl>