Pass around more info between commit visitors
Previously the visitors only passed around whether or not they needed
further history for processing in the forward stage of processing git
histories. This lets us pass around other information, e.g. if a
changelog file is removed, so that its contents can be disregarded when
processing the respective parent commit that introduced it.

Related: #210

Signed-off-by: Nils Philippsen <nils@redhat.com>