Enable longer changelog entries
Previously, only the git commit log subject would be considered for
changelog generation, which is quite restrictive because it's usually
limited to 50 characters in length. Also, people might want one commit
to contain more than one change and have one item for each.

This change enables both:

- By prepending an ellipsis (three dots '...' or the equivalent Unicode
  character '…') to the body, the subsequent text will be appended to
  the first item of a changelog entry.
- Dashed list items following the first item will be added to the
  changelog entry.

For example, this commit log text ...:

--- 8< ---
This is the first item but I'm feeling verbose

... today so it doesn't fit into the subject.
- And here I describe another change.
--- >8 ---

... would result in a changelog entry like this:

--- 8< ---
* Fri Mar 04 2022 Nils Philippsen <nils@redhat.com> - 1.0-1
- This is the first item but I'm feeling verbose today so it doesn't fit
  into the subject.
- And here I describe another change.
--- >8 ---

Additionally, it quotes percent signs in generated changelog entries
(but not in the `changelog` file, if it exists it will be included
verbatim) so RPM doesn't attempt to expand them as macros.

Fixes: #189, #234

Signed-off-by: Nils Philippsen <nils@redhat.com>