Always bump release if spec file exists
This bumps the release if can’t be parsed which may not have been the
case always. If the spec file doesn’t exist, nothing can be built from
the repo, so the release doesn’t have to be bumped. This is important so
the first commit(s) in a repo without a spec file are not counted (which
wouldn’t be intuitive).

Fixes: #247

Signed-off-by: Nils Philippsen <nils@redhat.com>