doc: use the appropriate lexers for code blocks
The default in rst is python, which results in every "with" and "in"
being highlighted, which isn't particularly useful e.g. for changelogs.
So specify 'text' as the default, and override this with 'spec' or
'bash' in the few cases which are not plain text.

(It would be nice if rst had a highlighter for changelogs, but it seems
it doesn't.)

Signed-off-by: Zbigniew Jędrzejewski-Szmek <zbyszek@in.waw.pl>