Change %autochangelog formatting to use dash before version-release
Packaging Guidelines allow both styles (and also a third one where
version-release is specified in a separate line, but this isn't widely used).
But I think the style with dash is clearer and generally preferred:
- rpmdev-bumpspec and other tools use this
- most packagers do also:

$ rg --color=always -A1 '%changelog'| rg -v '%changelog|%autochangelog|^--$' *.spec >/tmp/all
$ rg -c '.' /tmp/all
20920
$ rg -c '> \d' /tmp/all
96
$ rg -c '> - \d' /tmp/all
20155

I'd like to see this changed because it's what most people expect and also
because mixed changes (with the manual changelog below autogenerated entries)
look inconsistent.

Signed-off-by: Zbigniew Jędrzejewski-Szmek <zbyszek@in.waw.pl>