convert: make commit message flexible and print info after conversion
The commit message is changed to 'Convert to %autorelease' /
'Convert to %autochangelog' / 'Convert to %autorelease and %autochangelog'
depending on options and what is already in the spec file, and at the end
'Converted to %autorelease and %autochangelog, and committed to git'
(or some subset if less work was done) is printed.

I heard the complaint that people called 'convert' and weren't sure what
exactly happened. This should help.

Signed-off-by: Zbigniew Jędrzejewski-Szmek <zbyszek@in.waw.pl>