Drop skipping of changelog entries based on filename patterns
The basic premise is that koji can build any dist-git commit in history (*).
Thus trying to skip some commits in changelog based on whether only
"unimportant" files were touched is misguided and can never work correctly (**).
If the packager makes a commit that shouldn't contribute to changelog, they
should use "[skip changelog]". Drop this whole chunk of logic.

Fixes #239.

(*) If %autorelease is used, this is obvious. Empty commits are used for
rebuilds. Without %autorelease, this is less ovious, but is true nonetheless. We
know that koji will reject builds with the same nevra, so if the spec file is
not changed, a repeated build usually will be unsuccessful. But this doesn't
take into account %{dist} — if we wait long enough, we'll get a different
%{dist} string after a new release, and then we can build the package. But even
ignoring that case, if we have a few commits with the same spec contents, we
expect that only one of them could have been built successfully, but we don't
know _which_ one, so we have no choice but to include every one in the
changelog. So effectively we can treat the cases with and without %autorelease
as equivalent here.

(**) Some examples:
- as in #239: first commit changes the spec file but omits sources, second
  commit updates sources. The first commit needs to be in the changelog because
  there's a new version, the second commit needs to be in the changelog so that
  -2 version corresponds to the build nevra used by koji.

- we have a commit which only changes sources with the title "Upload the
  correct tarball". In this case, -1 was built successfully but is wrong, and
  -2 also builds successfully but is quite different, so both need to be in the
  changelog.

Signed-off-by: Zbigniew Jędrzejewski-Szmek <zbyszek@in.waw.pl>