Drop compatibility with pygit < 1.4
Currently, rpmautospec is available on current Fedora versions as well
as RHEL/CS >= 9, these all have a newer version. The plan for the future
is to make it available on RHEL/CS 8, too, on top of the Python 3.11
stack, which will need a backport of a current pygit2 version as the one
available (0.26) is much too old.

Signed-off-by: Nils Philippsen <nils@redhat.com>