Improve coverage of _get_rpmverflags() unit tests
Make parsing the abridged spec file fail in relevant tests and don’t
rely on rpm (>= 4.20) to do it, and just tell coverage that the loop is
unlikely to finish without being broken out of.

This is a more straight-forward solution than PR#182, i.e. commit
a5a2653498522d6f6255d522cb232b42411ebdea.

Also, explicitly declare `rpm` a third-party module. This comes to the
aid of `ruff check` which confuses it with a system library (why ever).

Signed-off-by: Nils Philippsen <nils@redhat.com>