Use python stdlib for date printing
Drops the use of the heavyweight Babel library for this single task. The
result is not thread-safe, but librpm probably isn't either and
rpmautospec is usually run in its own process.

This adds a simple context manager to changelog.py that allows us to set
the LC_TIME within a context block, so we don't forget to change it
back.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>