Make dependency on rpm package optional
Rpmautospec uses only a small subset of the rpm API, mostly working with
macros and parsing spec files. The minirpm package is a drop-in
replacement for just that functionality, it uses the native rpm*
libraries directly. It will get used if importing the official rpm
Python package fails.

Also, move the rpm macro file from `rpm/macros.d` to `rpm_macros.d`
because a directory called `rpm` in the project toplevel messes with
running the test suite.

Fixes: #256

Signed-off-by: Nils Philippsen <nils@redhat.com>